rename ingress due to module conflict
