+++ python/ray/serve/tests/test_pipeline_dag.py
@@ -10,7 +10,7 @@ from ray import serve
 from ray.serve.api import RayServeDAGHandle
 from ray.experimental.dag.input_node import InputNode
 from ray.serve.pipeline.api import build as pipeline_build
-from ray.serve.ingress import DAGDriver
+from ray.serve.driver import DAGDriver
 
 RayHandleLike = TypeVar("RayHandleLike")
 NESTED_HANDLE_KEY = "nested_handle"
